rename mod sequential to mod container
